Generate shutdown confirmation dialog
When the system firmware update includes MRC training update, it can
cause the system to boot slower after the update is applied. Currently a
user notification is generated regarding the slow boot. This CL
generates an additional shutdown confirmation dialog to the user
highlighting the same. Here is the screenshot capturing the confirmation
dialog:
https://screenshot.googleplex.com/8przUKEDrRz
Pressing "Continue" button in the screenshot will continue with the
restart. Pressing "Do it later" button will cancel the restart. Also
added a test case to UpdateNotificationControllerTest to verify the
same. Ensured that all the tests in that test suite ran successfully.

Bug: b/129056373
Change-Id: Iba3b53dffad9216e1e18a94fc80328d2cbdae4fc
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2001740
Commit-Queue: Karthikeyan Ramasubramanian <kramasub@chromium.org>
Reviewed-by: James Cook <jamescook@chromium.org>
Reviewed-by: Jenny Zhang <jennyz@chromium.org>
Cr-Commit-Position: refs/heads/master@{#736678}