[Autofill Auth] Cancelling PaymentsRequest if user cancels pending dialog.
Currently when the user cancels the Verifying Pending Dialog immediately after
card selection, the Payments Request to fetch unmasking details is not
cancelled. This means that when the request is responded to, the CVC dialog is
shown.

However, cancelling the Verifying Pending Dialog should fallback to a regular
CVC auth flow.

This CL fixes both these issues.

Bug: 949269
Change-Id: If2dc973c5d938cf05d43694342894b572205cbe7
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2028116
Commit-Queue: Manas Verma <manasverma@google.com>
Reviewed-by: Jared Saul <jsaul@google.com>
Cr-Commit-Position: refs/heads/master@{#736700}