Move live registration lookup from ServiceWorkerStorage to ServiceWorkerRegistry
Before this CL, FindRegistrationForId() worked like the below:
1. Check if storage is initialized/disabled.
   * if not initialized, perform lazy initialization then go back
   step 1.
   * if disabled, abort registration lookup.
2. Check if there are stored registrations for the given origin.
   * if NOT exists, only checks installing registrations.
3. Check if there is a live registration for the given registration id.
   * if exists, return the found registration.
4. Check database.

This ordering doesn't fit well for the Storage Service migration
because step 1, 2 and 4 should be performed in the storage service
(in the future) but step 3 should remain in content/. Keeping this
ordering will require an additional IPC round-trip which is undesirable.

This CL changes the above steps like the below so that we won't need
extra IPC:
0. Introduce a new field in ServiceWorkerRegistration which tells that
   if the registration is stored in persistent storage.
1. Check if storage is disabled. If disabled, abort registration
   lookup.
2. Check if there is a live registration for the given registration id.
   * if exists and the registration is stored on storage or in
     installing state, return the found registration.
3. Initialize storage if needed.
4. Check if there are stored registrations for the given origin.
   * This is the same as the step 2 of the previous behavior.
5. Check database.

The new behavior should be almost the same as long as we maintain the
field which indicates if a registration is stored accordingly. This CL
makes sure the condition by updating the field when a registration is
created from ServiceWorkerDatabase::RegistrationData or
ServiceWorkerRegistry::DidStoreRegistration() is called.

Bug: 1039200
Change-Id: I9e351c04a2eaa516a75274757778f604eadf951a
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2021783
Commit-Queue: Kenichi Ishibashi <bashi@chromium.org>
Reviewed-by: Hiroki Nakagawa <nhiroki@chromium.org>
Cr-Commit-Position: refs/heads/master@{#736743}