media/gpu: Replace VLOG(1)s with LOG(ERROR) in V4L2(S)/VaapiVDAs
VDAs used on ChromeOS logs real errors with VLOG(1). A verbose
level in media/gpu is not set, so that these logs are not left
in chrome log. This is hard for us to figure out issues from
user reports.

This replaces VLOG(1) for real errors with LOG(ERROR) in
VDAs used on ChromeOS.

Bug: b:146518063
Test: vda_test on kevin and hatch
Change-Id: Ic2e7b3da9e5aba5ab6dd582888b7c0a0b6d29a97
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2027010
Commit-Queue: Hirokazu Honda <hiroh@chromium.org>
Reviewed-by: Alexandre Courbot <acourbot@chromium.org>
Cr-Commit-Position: refs/heads/master@{#736776}