[signin] Support Dice in extension consent flow
Before this change, any account added during the extension consent flow
was not going through Dice, and thus was not properly added to Chrome.
This prevented the account to be properly managed, and in particular it
was not possible to refresh the access tokens.

This CL attaches Dice headers to requests coming from the extension
consent flow, and the account is now added to Chrome.

Other guest web views (such as the InlineLoginUI) are still ignored by
Dice. It is unclear whether they really need to be ignored, but to
minimize the change, this CL doesn't modify this behavior.

Change-Id: I869ab5e699b7ee6cb556c17555be502893e3ea67
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1954387
Commit-Queue: David Roger <droger@chromium.org>
Reviewed-by: Alex Ilin <alexilin@chromium.org>
Reviewed-by: Reilly Grant <reillyg@chromium.org>
Reviewed-by: Alex Moshchuk <alexmos@chromium.org>
Cr-Commit-Position: refs/heads/master@{#736813}