Added histograms around the `missing permission` Android infobar
Records the amount of times the infobar is shown and how many it times
it does not need to be shown.
Records the action that was taken as a result (granted/dismissed/denied)

This helps us prepare some data for when the Android-level ephemerality
changes will land and give us a baseline to compare to as we expect the
number of times this infobar is shown to increase significantly.

Bug: 1047184
Change-Id: If9dbec366d8c57b37585b990531432f5a1235ae7
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2030865
Commit-Queue: Andy Paicu <andypaicu@chromium.org>
Commit-Queue: Kamila Hasanbega <hkamila@chromium.org>
Reviewed-by: Kamila Hasanbega <hkamila@chromium.org>
Reviewed-by: Brian White <bcwhite@chromium.org>
Auto-Submit: Andy Paicu <andypaicu@chromium.org>
Cr-Commit-Position: refs/heads/master@{#736910}