[Autofill Assistant] Make edit button fully removable.
Before this change, there was already a boolean attribute to let the
layout know it has an edit button. It has not been in use however.
This CL reads the attribute (defaulting to true, for backwards
compatibility) and disables the third column in the layout if necessary.

This CL also removes the |can_add_items| attribute that seems to have
been replaced with |add_button_text|.

Screenshot (after):
https://screenshot.googleplex.com/eKQcLrJBey9

Bug: b/146041576
Change-Id: I5865b71c7e9a726d1793b3a0016b42bebb4c45fe
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2027454
Reviewed-by: Sandro Maggi <sandromaggi@google.com>
Reviewed-by: Clemens Arbesser <arbesser@google.com>
Commit-Queue: Sandro Maggi <sandromaggi@google.com>
Cr-Commit-Position: refs/heads/master@{#736914}