Convert test files in c/b/r/c/a/chromevox to ES6 classes
This is the continuing work to modernize accessibility extension js and make them uniform.

The test js in Chrome is somewhat complicated because it needs to run through V8 shell in a first pass and in a real Chrome renderer when actually running the test.

The first pass does not have access to the bulk of the test-only modules.

Steps for conversion:
- run automated conversion script
- change classes from
class Foo
to
Foo = class
(assumes global scope)
- move 'first pass' data onto prototype; unfortunately, we cannot instantiate the test object class in the V8 shell
- remove all empty super() constructors (since default super constructors are automatically called)

Test: all tests pass.

Enforce object-shorthand in c/b/r/c/accessibility

Change-Id: I7cf69c4bdcaf7c71d7d0355e802a2e48086fd9a8
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2025631
Commit-Queue: David Tseng <dtseng@chromium.org>
Reviewed-by: Dominic Mazzoni <dmazzoni@chromium.org>
Cr-Commit-Position: refs/heads/master@{#736918}