A11y: Add crash-key to track down AXTree crash
The linked bug is a crash caused by a dangling pointer; we hypothesize
that it occurs in one of the few places where AXTree::Unserialize is
called without an error being treated as fatal.

This change adds a CHECK around one such call (which didn't have any
error handling), and logs some data with a crash-key in a second case
(which had existing error handling, but didn't treat it as fatal). This
will help us prove (or disprove) these scenarios as the culprit.

Bug: 1034755
Change-Id: I1b1164295b25e508c61ef97c1ea0ce093b52ba18
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2025836
Reviewed-by: Scott Violet <sky@chromium.org>
Reviewed-by: Robert Sesek <rsesek@chromium.org>
Reviewed-by: Ethan Jimenez <ethavar@microsoft.com>
Reviewed-by: Aaron Leventhal <aleventhal@chromium.org>
Commit-Queue: Ian Prest <iapres@microsoft.com>
Cr-Commit-Position: refs/heads/master@{#736955}