Add match type support for RTLookupResponse.
Several changes in realtimeapi.proto are made in this CL:

1. The old cache_expression field is renamed to
cache_expression_covering_match, this is to avoid breaking
old clients that are using this field.

2. Add a new enum named cache_expression_match_type,
if this field is not set, ignore this entry; if this field
is set to EXACT_MATCH, this entry is applicable for exact
url matching; if this field is set to COVERING_MATCH, this
entry is applicable for host-suffix path-prefix matching.

3. Add a new cache_expression field, this field will be
used as the cache expression for new clients.

For details, see go/chrome-protego-cache-matching-field.
Server-side change: http://cl/292004484

Bug: 1041675
Change-Id: I5a40babdb9be0ebad8661361a93194daa445f479
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2008080
Reviewed-by: Varun Khaneja <vakh@chromium.org>
Commit-Queue: Xinghui Lu <xinghuilu@chromium.org>
Cr-Commit-Position: refs/heads/master@{#736972}