autotest private api: Allow setting fullscreen & minimized in tablet.
Ran into this issue when trying to write a tast-test. The main way is
to press the home button which we can use an accelerator for, but its
still helpful for tests to do this, and you can still you minimize and
fullscreen accelerators in tablet mode with a keyboard.

Added a test and formatted the file.

Test: added test, manual
Bug: 1045185
Change-Id: Iead7516d133929b93452bf4f29e24e462e728a8b
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2025967
Commit-Queue: Sammie Quon <sammiequon@chromium.org>
Reviewed-by: Mitsuru Oshima (slow in TOK) <oshima@chromium.org>
Cr-Commit-Position: refs/heads/master@{#736980}