clang upload script: Move list of bots from docs to Cq-Include-Trybots:.
This way, the cq makes sure that we run all the bots we want to run
(as long as we use the upload script), which is less error-prone.

This doesn't work for internal chrome/try bots yet, so these stay on
the old system for now.

Bug: 1047347
Change-Id: Iedecf56d4cfa5f6e923b716b927147ac3c28dedc
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2016844
Auto-Submit: Nico Weber <thakis@chromium.org>
Reviewed-by: Hans Wennborg <hans@chromium.org>
Commit-Queue: Hans Wennborg <hans@chromium.org>
Cr-Commit-Position: refs/heads/master@{#736982}