Update IsInsecureCapturingInProgress()
IsInsecureCapturingInProgress() will now report capturing in progress
if the whole desktop, the window for Chrome, or the specific tab is
being captured. Adds UpdateTarget() to specify the actual target being
captured, so that the code can check if the tab this is being called
for is actually the target, not just the tab doing the recording.
Default target is still the tab doing the recording.

Also updates Clear Key CDM so that every response to
QueryOutputProtectionStatus() generates a 'keychange' event for
a dummy key ID. If capturing reported, then the key status is
'output-protected'. If no capturing reported, key status is 'usable'.

Adds additional output protection tests that use the fake UI to select
capturing of the monitor, window, or browser window. Previous tests
ran with kUseFakeUIForMediaStream not specified, which defaulted to
"any".

Bug: 856276
Test: new EME browser_tests pass, capturing tested manually as well
Change-Id: I50df86380cfc115a218d92ba54027294f6bb3b54
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1987271
Commit-Queue: John Rummell <jrummell@chromium.org>
Reviewed-by: Yuri Wiitala <miu@chromium.org>
Reviewed-by: Xiaohan Wang <xhwang@chromium.org>
Cr-Commit-Position: refs/heads/master@{#737054}