Fix hanging UI on invalid Assistant query.
In |AssistantInteractionController::OnOpenAppResponse| we forgot to call
the callback if it is invoked outside of a valid conversation.
This blocks the entire |AssistantInteractionSubscriber| mojom pipeline,
causing the Assistant UI to become unresponsive.

The fix was easy (call the callback), but I used this opportunity to add
the first unittests for the |AssistantInteractionController|.

Bug: b/146227377
Change-Id: I705aed3d4cabf3761de339d7d1c8a1ce6178b90d
Tests: New ash_unittests AssistantInteractionController*
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2029607
Reviewed-by: Xiyuan Xia <xiyuan@chromium.org>
Reviewed-by: Xiaohui Chen <xiaohuic@chromium.org>
Commit-Queue: Jeroen Dhollander <jeroendh@google.com>
Cr-Commit-Position: refs/heads/master@{#737058}